fix: Correctly parse v4 messages for detection
The `MessageDetectionService` was failing to detect encrypted messages because the `MessageParser` did not support the `~BCEv4~` message format.

This change fixes the bug by:
1.  Moving the shared data classes (`TinkMessage`, `DecryptedMessage`) to a common `crypto/model` package.
2.  Centralizing the v4 message parsing logic into a new public method in `MessageParser`.
3.  Updating `MessageParser`'s `getBarcodeNameFromMessage` method to correctly handle the v4 format.
4.  Refactoring `EncryptionManager` to use the centralized parser.

This change should resolve the issue where encrypted messages were not being detected and highlighted.
