+++ app/src/main/java/com/hereliesaz/barcodencrypt/BarcodeApplication.kt
@@ -2,7 +2,6 @@ package com.hereliesaz.barcodencrypt
 
 import android.app.Application
 import com.google.crypto.tink.aead.AeadConfig
-import com.google.crypto.tink.hybrid.HybridConfig
 import java.security.GeneralSecurityException
 
 class BarcodeApplication : Application() {
@@ -11,9 +10,8 @@ class BarcodeApplication : Application() {
         super.onCreate()
         try {
             AeadConfig.register()
-            HybridConfig.register() // For X25519 and other hybrid encryption schemes
         } catch (e: GeneralSecurityException) {
-            throw RuntimeException("Tink initialization failed", e)
+            throw RuntimeException(e)
         }
     }
 }
