--- app/src/main/java/com/hereliesaz/barcodencrypt/crypto/CryptoScheme.kt
@@ -1,19 +1,0 @@
-package com.hereliesaz.barcodencrypt.crypto
-
-import com.hereliesaz.barcodencrypt.crypto.model.DecryptedMessage
-
-internal interface CryptoScheme {
-    fun encrypt(
-        plaintext: String,
-        ikm: String,
-        keyName: String,
-        counter: Long,
-        options: List<String> = emptyList(),
-        maxAttempts: Int = 0
-    ): String?
-
-    fun decrypt(
-        ciphertext: String,
-        ikm: String
-    ): DecryptedMessage?
-}
