+++ app/src/main/java/com/hereliesaz/barcodencrypt/crypto/model/TinkMessage.kt
@@ -1,5 +1,7 @@
 package com.hereliesaz.barcodencrypt.crypto.model
 
+// This class is used by both the crypto and util packages, so it's defined here.
+// It's internal to the module.
 internal data class TinkMessage(
     val salt: ByteArray,
     val ciphertext: ByteArray,
