+++ app/src/main/java/com/hereliesaz/barcodencrypt/services/OverlayService.kt
@@ -45,7 +45,9 @@ import com.hereliesaz.barcodencrypt.ui.PasswordDialog
 import com.hereliesaz.barcodencrypt.ui.PasswordScannerTrampolineActivity
 import com.hereliesaz.barcodencrypt.ui.theme.BarcodencryptTheme
 import com.hereliesaz.barcodencrypt.ui.theme.DisabledRed
-import com.hereliesaz.barcodencrypt.util.*
+import com.hereliesaz.barcodencrypt.util.Constants
+import com.hereliesaz.barcodencrypt.util.MessageParser
+import com.hereliesaz.barcodencrypt.util.PasswordPasteManager
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.IntentFilter
@@ -66,7 +68,6 @@ class OverlayService : Service(), LifecycleOwner, ViewModelStoreOwner, SavedStat
     private val overlayState = mutableStateOf<OverlayState>(OverlayState.Initial)
     private var encryptedText: String? = null
     private var barcodeName: String? = null
-    private var passwordFieldId: String? = null
     private val scannedSequence = mutableListOf<String>()
     private var scanReceiver: BroadcastReceiver? = null
 
@@ -97,11 +98,7 @@ class OverlayService : Service(), LifecycleOwner, ViewModelStoreOwner, SavedStat
             override fun onReceive(context: Context?, intent: Intent?) {
                 if (intent?.action == ACTION_SCAN_RESULT) {
                     val scannedValue = intent.getStringExtra(Constants.IntentKeys.SCAN_RESULT)
-                    if (passwordFieldId != null && scannedValue != null) {
-                        handlePasswordUnlock(scannedValue)
-                    } else {
-                        handleScannedKey(scannedValue)
-                    }
+                    handleScannedKey(scannedValue)
                 }
             }
         }
@@ -109,24 +106,6 @@ class OverlayService : Service(), LifecycleOwner, ViewModelStoreOwner, SavedStat
             registerReceiver(scanReceiver, IntentFilter(ACTION_SCAN_RESULT), RECEIVER_NOT_EXPORTED)
         } else {
             registerReceiver(scanReceiver, IntentFilter(ACTION_SCAN_RESULT))
-        }
-    }
-
-    private fun handlePasswordUnlock(scannedBarcodeValue: String) {
-        val fieldId = passwordFieldId ?: return
-        serviceScope.launch(Dispatchers.IO) {
-            val dao = AppDatabase.getDatabase(application).passwordEntryDao()
-            val passwordManager = PasswordManager(dao)
-            val unlockedPassword = passwordManager.unlockPassword(fieldId, scannedBarcodeValue)
-
-            withContext(Dispatchers.Main) {
-                if (unlockedPassword != null) {
-                    PasswordPasteManager.paste(unlockedPassword)
-                    overlayState.value = OverlayState.Success("Password unlocked and pasted!")
-                } else {
-                    overlayState.value = OverlayState.Failure
-                }
-            }
         }
     }
 
@@ -173,7 +152,6 @@ class OverlayService : Service(), LifecycleOwner, ViewModelStoreOwner, SavedStat
                     stopSelf()
                     return START_NOT_STICKY
                 }
-                passwordFieldId = intent.getStringExtra(Constants.IntentKeys.PASSWORD_FIELD_ID)
                 overlayState.value = OverlayState.PasswordIcon
                 createOverlay(bounds)
             }
