+++ app/src/main/java/com/hereliesaz/barcodencrypt/ui/SettingsActivity.kt
@@ -10,7 +10,6 @@ import android.provider.Settings
 import android.view.autofill.AutofillManager
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -23,13 +22,11 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.compose.LocalLifecycleOwner
-import androidx.lifecycle.lifecycleScope
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.annotation.RequiresApi
-import kotlinx.coroutines.launch
 import com.hereliesaz.barcodencrypt.MainActivity
 import com.hereliesaz.barcodencrypt.R
 import com.hereliesaz.barcodencrypt.services.BarcodeAutofillService
@@ -43,22 +40,6 @@ class SettingsActivity : ComponentActivity() {
 
     private val viewModel: SettingsViewModel by viewModels {
         SettingsViewModelFactory(applicationContext)
-    }
-
-    private val backupLauncher = registerForActivityResult(ActivityResultContracts.CreateDocument("application/octet-stream")) { uri ->
-        uri?.let {
-            lifecycleScope.launch {
-                com.hereliesaz.barcodencrypt.util.BackupManager.backupDatabase(this@SettingsActivity, it)
-            }
-        }
-    }
-
-    private val restoreLauncher = registerForActivityResult(ActivityResultContracts.GetContent()) { uri ->
-        uri?.let {
-            lifecycleScope.launch {
-                com.hereliesaz.barcodencrypt.util.BackupManager.restoreDatabase(this@SettingsActivity, it)
-            }
-        }
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -87,11 +68,7 @@ class SettingsActivity : ComponentActivity() {
                         })
                     },
                     screenContent = {
-                        SettingsScreen(
-                            viewModel = viewModel,
-                            onBackup = { backupLauncher.launch("barcodencrypt_backup.db") },
-                            onRestore = { restoreLauncher.launch("*/*") }
-                        )
+                        SettingsScreen(viewModel)
                     }
                 )
             }
@@ -127,11 +104,7 @@ class SettingsActivity : ComponentActivity() {
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
-fun SettingsScreen(
-    viewModel: SettingsViewModel,
-    onBackup: () -> Unit,
-    onRestore: () -> Unit
-) {
+fun SettingsScreen(viewModel: SettingsViewModel) {
     val context = LocalContext.current
     val lifecycleOwner = LocalLifecycleOwner.current
     val autofillManager = remember { context.getSystemService(AutofillManager::class.java) }
@@ -197,16 +170,6 @@ fun SettingsScreen(
 
         Button(onClick = { viewModel.logout() }) {
             Text("Log out")
-        }
-
-        Spacer(modifier = Modifier.height(32.dp))
-
-        Button(onClick = onBackup) {
-            Text("Backup Database")
-        }
-        Spacer(modifier = Modifier.height(16.dp))
-        Button(onClick = onRestore) {
-            Text("Restore Database")
         }
     }
 }
