+++ app/src/main/java/com/hereliesaz/barcodencrypt/util/Constants.kt
@@ -17,6 +17,5 @@ object Constants {
         const val CORRECT_KEY = "com.hereliesaz.barcodencrypt.CORRECT_KEY"
         const val BOUNDS = "com.hereliesaz.barcodencrypt.BOUNDS"
         const val TUTORIAL_BARCODE = "com.hereliesaz.barcodencrypt.TUTORIAL_BARCODE"
-        const val PASSWORD_FIELD_ID = "com.hereliesaz.barcodencrypt.PASSWORD_FIELD_ID"
     }
 }
