+++ app/src/main/java/com/hereliesaz/barcodencrypt/util/MessageParser.kt
@@ -1,10 +1,29 @@
 package com.hereliesaz.barcodencrypt.util
 
+import com.google.gson.Gson
+import com.hereliesaz.barcodencrypt.crypto.model.TinkMessage
+
 object MessageParser {
+
+    private const val HEADER_PREFIX_V4 = "~BCEv4~"
+
+    fun parseV4Message(ciphertext: String): TinkMessage? {
+        if (!ciphertext.startsWith(HEADER_PREFIX_V4)) {
+            return null
+        }
+        return try {
+            val json = String(android.util.Base64.decode(ciphertext.removePrefix(HEADER_PREFIX_V4), android.util.Base64.NO_WRAP), Charsets.UTF_8)
+            val gson = Gson()
+            gson.fromJson(json, TinkMessage::class.java)
+        } catch (e: Exception) {
+            null
+        }
+    }
 
     fun getBarcodeNameFromMessage(message: String): String? {
         val parts = message.split("::")
         return when {
+            message.startsWith(HEADER_PREFIX_V4) -> parseV4Message(message)?.keyName
             message.startsWith("BCE::v2") && parts.size >= 4 -> parts[3]
             message.startsWith("~BCE~") -> {
                 try {
